Define Incorrect Behavior of haveInvoiceItems (#871)
* Define Incorrect Behavior of haveInvoiceItems

- Define incorrect behavior of `InventoryBatch.haveInvoiceItems()` in
  model unit test

`haveInvoiceItems` returns `true` is `invoiceItems` is empty. This
should likely be revered or method name should change.

* Fix InventoryBatch.haveInventoryItems()

- Fix InventoryBatch.haveInventoryItems() to return true when not empty

* Fix conditional validation in inventory-batch

- Fix conditional validation in inventory-batch to fix of
  `haveInvoiceItems`.
- Add tests for validations

As 'haveInvoiceItems` method has been flipped to now return true
when `invoiceItems` are empty the validations for the model were made
incorrect. This commit adjusts those validations to align with the
updated method.